+++ crates/client/sync/src/l2.rs
@@ -18,6 +18,7 @@ use dp_felt::FeltWrapper;
 use futures::{stream, StreamExt};
 use num_traits::FromPrimitive;
 use starknet_api::hash::StarkFelt;
+use starknet_api::transaction::TransactionHash;
 use starknet_core::types::StateUpdate;
 use starknet_providers::sequencer::models::StateUpdateWithBlock;
 use starknet_providers::{ProviderError, SequencerGatewayProvider};
@@ -55,12 +56,16 @@ pub struct L2StateUpdate {
     pub block_hash: Felt,
 }
 
-fn store_new_block(backend: &DeoxysBackend, block: &DeoxysBlock) -> Result<(), DeoxysStorageError> {
+fn store_new_block(
+    backend: &DeoxysBackend,
+    block: &DeoxysBlock,
+    reverted_txs: &[TransactionHash],
+) -> Result<(), DeoxysStorageError> {
     let sw = PerfStopwatch::new();
     let mut tx = WriteBatchWithTransaction::default();
 
     let mapping = backend.mapping();
-    mapping.write_new_block(&mut tx, block)?;
+    mapping.write_new_block(&mut tx, block, reverted_txs)?;
 
     let db_access = backend.expose_db();
     db_access.write(tx).map_err(MappingDbError::from)?;
@@ -80,7 +85,7 @@ async fn l2_verify_and_apply_task(
     sync_timer: Arc<Mutex<Option<Instant>>>,
     telemetry: TelemetryHandle,
 ) -> anyhow::Result<()> {
-    while let Some(L2ConvertedBlockAndUpdates { converted_block, state_update, class_update }) =
+    while let Some(L2ConvertedBlockAndUpdates { converted_block, reverted_txs, state_update, class_update }) =
         channel_wait_or_graceful_shutdown(pin!(updates_receiver.recv())).await
     {
         let block_n = converted_block.block_n();
@@ -124,7 +129,7 @@ async fn l2_verify_and_apply_task(
         let storage_diffs = state_update.state_diff.storage_diffs.clone();
 
         let (r1, (r2, (r3, r4))) = rayon::join(
-            || store_new_block(&backend, &converted_block),
+            || store_new_block(&backend, &converted_block, &reverted_txs),
             || {
                 rayon::join(
                     || store_state_update(&backend, block_n, state_update),
@@ -185,6 +190,7 @@ async fn l2_verify_and_apply_task(
 
 pub struct L2ConvertedBlockAndUpdates {
     pub converted_block: DeoxysBlock,
+    pub reverted_txs: Vec<TransactionHash>,
     pub state_update: StateUpdate,
     pub class_update: Vec<ContractClassData>,
 }
@@ -204,7 +210,12 @@ async fn l2_block_conversion_task(
                         let sw = PerfStopwatch::new();
                         let converted_block = convert_block(block, chain_id).context("converting block")?;
                         stopwatch_end!(sw, "convert_block: {:?}");
-                        anyhow::Ok(L2ConvertedBlockAndUpdates { converted_block, state_update, class_update })
+                        anyhow::Ok(L2ConvertedBlockAndUpdates {
+                            converted_block: converted_block.block,
+                            reverted_txs: converted_block.reverted_txs,
+                            state_update,
+                            class_update,
+                        })
                     }),
                     (updates_recv, chain_id),
                 )
@@ -266,7 +277,7 @@ async fn l2_pending_block_task(
                 .context("converting pending block")?;
             let storage_update = crate::convert::state_update(state_update);
             storage
-                .write_pending(&mut tx, &block, &storage_update)
+                .write_pending(&mut tx, &block.block, &storage_update)
                 .context("writing pending to rocksdb transaction")?;
         } else {
             storage.write_no_pending(&mut tx).context("writing no pending to rocksdb transaction")?;